requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
openai==1.58.1
pydantic==2.6.1